Added files for the HomeScreen + Nav
Added all files needed for the HomeScreen and the the NavigationBar. The home.dart file is where the assessment should be linked to, while the main_page displays what the "home" should display. I did not include the main file as I think we should use Saif's for now. Also included my widget for large text on the main_page